fix all imports for plotting.py
